% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopPropagHate.R
\name{stopPropagHate}
\alias{stopPropagHate}
\title{Stop PropagHate function}
\usage{
stopPropagHate(data_frame, hate_type, language)
}
\arguments{
\item{data_frame, }{is a data.frame containing a column named "text".}

\item{hate_type, }{is the type of hate you want to measure. It can be general "hate" speech, "sexism" or "racism".}

\item{language, }{we provide this funtion both for Enlish "en" and Portuguese "pt".}
}
\description{
This function allows you to evaluate if a set of text messages constains hate speech.
}
\examples{
texts_data_frame <- data.frame(id = c(1,2,3), text = c("Lugar de mulher e na cozinha, isto e a verdade", "gorda e feia", "mais uma mensagem de teste")) 
stopPropagHate(texts_data_frame, "sexism", "pt")
}
\keyword{hate}
\keyword{racism,}
\keyword{sexism}
\keyword{speech,}
