% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopPropagHate.R
\name{stopPropagHateMessage}
\alias{stopPropagHateMessage}
\title{Stop PropagHate function}
\usage{
stopPropagHateMessage(message, hate, language)
}
\arguments{
\item{message, }{is a string with the text to be classified.}

\item{language, }{we provide this funtion both for Enlish "en" and Portuguese "pt".}

\item{hate_type, }{is the type of hate you want to measure. It can be general "hate" speech, "sexism" or "racism".}
}
\description{
This function allows you to evaluate if a set of text messages constains hate speech. The usage of the function stopPropagHate requires the download of tokenizer and model files to the working directory. For an automatic download and usage without prompt check the function stopPropagHateWithoutPrompt.
}
\examples{
message <- "Call me sexist but female sports anchors r the worst. Makes me uncomfortable when women know more about sports than I do", "May Allah bless him with 72 virgin pigs."
stopPropagHateMessage(message, "hate", "en")
}
\keyword{hate}
\keyword{racism,}
\keyword{sexism}
\keyword{speech,}
