% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopPropagHate.R
\name{test_package}
\alias{test_package}
\title{test_package function}
\usage{
test_package()
}
\description{
This function allows you to evaluate if a set of text messages constains hate speech. The usage of the function stopPropagHate requires the download of tokenizer and model files to the working directory. For an automatic download and usage without prompt check the function stopPropagHateWithoutPrompt.
}
\examples{
test_package()
}
\keyword{hate}
\keyword{racism,}
\keyword{sexism}
\keyword{speech,}
